pyproject.toml
[project]
name = "lessons"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleksandr Nazarevych",email = "oleksandr.o.nazarevych@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "polygon-api-client (==1.14.5)",
    "pandas (>=2.3.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "vectorbt (>=0.28.0,<0.29.0)",
    "yfinance (>=0.2.65,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
